@@ -3,77 +3,74 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 27
-    defaultConfig {
-        applicationId "pl.tobzzo.miechowskiepowietrze"
-        minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 4
-        versionName "1.4.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+  compileSdkVersion 29
+  defaultConfig {
+    applicationId "pl.tobzzo.miechowskiepowietrze"
+    minSdkVersion 14
+    targetSdkVersion 29
+    versionCode 4
+    versionName "1.4.0"
+    testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+  }
+  buildTypes {
+    debug {
+      minifyEnabled false
+      proguardFiles 'proguard-rules.pro'
+
+      buildConfigField "String", "hiddenPassword1", "\"${hiddenPassword1}\""
+      buildConfigField "String", "hiddenPassword2", "\"${hiddenPassword2}\""
+
+      debuggable true
     }
-    buildTypes {
-        debug {
-            minifyEnabled false
-            proguardFiles  'proguard-rules.pro'
+    release {
 
-            buildConfigField "String", "hiddenPassword1", "\"${hiddenPassword1}\""
-            buildConfigField "String", "hiddenPassword2", "\"${hiddenPassword2}\""
+      minifyEnabled false
+      proguardFiles 'proguard-rules.pro'
 
-            debuggable true
-        }
-        release {
+      buildConfigField "String", "hiddenPassword1", "\"${hiddenPassword1}\""
+      buildConfigField "String", "hiddenPassword2", "\"${hiddenPassword2}\""
 
-            minifyEnabled false
-            proguardFiles  'proguard-rules.pro'
-
-            buildConfigField "String", "hiddenPassword1", "\"${hiddenPassword1}\""
-            buildConfigField "String", "hiddenPassword2", "\"${hiddenPassword2}\""
-
-            debuggable false
-        }
+      debuggable false
     }
+  }
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.daimajia.numberprogressbar:library:1.4@aar'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
-//    implementation 'com.koushikdutta.ion:ion:2.2.1'
-    implementation 'com.github.PhilJay:MPAndroidChart:v3.0.1'
-    implementation 'com.google.android.gms:play-services-analytics:16.0.8'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "com.android.support.constraint:constraint-layout:1.1.3"
-    implementation "com.android.support:support-fragment:27.1.1"
-    implementation "com.android.support:support-v4:27.1.1"
-    implementation "com.android.support:support-annotations:28.0.0"
-    implementation "com.android.support:design:27.1.1"
-    implementation "com.android.support:cardview-v7:27.1.1"
-    implementation 'com.google.dagger:dagger:2.21'
-    annotationProcessor 'com.google.dagger:dagger-compiler:2.21'
-    implementation 'com.google.dagger:dagger-android:2.21'
-    implementation 'com.google.dagger:dagger-android-support:2.21' // if you use the support libraries
-    implementation 'com.jakewharton.timber:timber:4.7.1'
-    annotationProcessor 'com.google.dagger:dagger-android-processor:2.21'
-    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.8'
-    implementation 'io.reactivex.rxjava2:rxkotlin:2.3.0'
-    // retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.5.0'
-    implementation 'com.squareup.retrofit2:converter-gson:2.1.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:3.12.1'
+  implementation fileTree(dir: 'libs', include: ['*.jar'])
+  implementation 'com.daimajia.numberprogressbar:library:1.4@aar'
+  implementation 'androidx.appcompat:appcompat:1.0.2'
+  implementation 'androidx.recyclerview:recyclerview:1.0.0'
+  implementation 'com.github.PhilJay:MPAndroidChart:v3.0.1'
+  implementation 'com.google.android.gms:play-services-analytics:17.0.0'
+  implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+  implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+  implementation 'androidx.fragment:fragment:1.0.0'
+  implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+  implementation 'androidx.annotation:annotation:1.1.0'
+  implementation 'com.google.android.material:material:1.0.0'
+  implementation 'androidx.cardview:cardview:1.0.0'
+  implementation 'com.google.dagger:dagger:2.24'
+  annotationProcessor "com.google.dagger:dagger-compiler:$dagger_version"
+  implementation "com.google.dagger:dagger-android:$dagger_version"
+  implementation "com.google.dagger:dagger-android-support:$dagger_version"
+  implementation 'com.jakewharton.timber:timber:4.7.1'
+  annotationProcessor "com.google.dagger:dagger-android-processor:$dagger_version"
+  implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+  implementation 'io.reactivex.rxjava2:rxjava:2.2.8'
+  implementation 'io.reactivex.rxjava2:rxkotlin:2.3.0'
+  // retrofit
+  implementation 'com.squareup.retrofit2:retrofit:2.5.0'
+  implementation 'com.squareup.retrofit2:converter-gson:2.1.0'
+  implementation 'com.squareup.okhttp3:logging-interceptor:3.12.1'
 
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-    testImplementation 'junit:junit:4.12'
-
+  androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
+    exclude group: 'com.android.support', module: 'support-annotations'
+  })
+  testImplementation 'junit:junit:4.12'
 }
 repositories {
-    mavenCentral()
-    maven { url "https://jitpack.io" }
+  mavenCentral()
+  maven { url "https://jitpack.io" }
 }
-
 
 apply plugin: 'com.google.gms.google-services'
