@@ -2,8 +2,8 @@ package pl.tobzzo.miechowskiepowietrze.activities
 
 import android.os.Bundle
 import android.os.Handler
-import android.support.v4.widget.SwipeRefreshLayout
-import android.support.v7.app.AppCompatActivity
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.ImageView
 import android.widget.ProgressBar
 import android.widget.ScrollView
